edit controller, remove trip location
